requirements.txt
# LOKTANTRA: The Sovereign Saga - Dependencies
# Google Cloud Platform
flask>=3.0,<4.0
gunicorn>=21.0,<23.0
google-cloud-firestore>=2.16,<3.0
google-cloud-aiplatform>=1.50,<2.0
google-cloud-logging>=3.10,<4.0
google-auth>=2.29,<3.0

# Utilities
python-dotenv>=1.0,<2.0
uuid6>=2024.1.12

# Testing
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
pytest-flask>=1.3,<2.0
